@@ -9,8 +9,8 @@ abstract class AppWebTestCase extends WebTestCase
     protected static function createKernel(array $options = array())
     {
         return new \VirtualKernel(
-            isset($options['environment']) ? $options['environment'] : 'test',
-            isset($options['debug']) ? $options['debug'] : true,
+            $options['environment'] ?? 'test',
+            $options['debug'] ?? true,
             'app'
         );
     }
